Task 3: refactor code according to eslint rules
