+++ src/app.js
@@ -1,8 +1,9 @@
-var express = require('express');
-var app = express();
-var db = require('./db');
-var user = require('./controllers/usercontroller')
-var game = require('./controllers/gamecontroller')
+const express = require('express');
+
+const app = express();
+const db = require('./db');
+const user = require('./controllers/usercontroller')
+const game = require('./controllers/gamecontroller')
 
 app.use(express.json());
 
@@ -11,6 +12,6 @@ db.sync();
 app.use('/api/auth', user);
 app.use('/api/game', game);
 
-app.listen(4000,function() {
+app.listen(4000,() => {
     console.log("App is listening on 4000");
 })
