+++ src/controllers/gamecontroller.js
@@ -1,20 +1,20 @@
-var router = require('express').Router();
-var Game = require('../models/game');
-var StatusCodes = require('http-status-codes').StatusCodes;
+const router = require('express').Router();
+const {StatusCodes} = require('http-status-codes');
+const Game = require('../models/game');
 
 router.use(require('../middleware/validate-session'))
 
 router.get('/all', (req, res) => {
     Game.findAll({ where: { owner_id: req.user.id } })
         .then(
-            function findSuccess(games) {
+            (games) => {
                 res.status(StatusCodes.OK).json({
-                    games: games,
+                    games,
                     message: "Data fetched."
                 })
             },
 
-            function findFail() {
+            () => {
                 res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
                     message: "Internal Server Error"
                 })
@@ -25,18 +25,20 @@ router.get('/all', (req, res) => {
 router.get('/:id', (req, res) => {
     Game.findOne({ where: { id: req.params.id, owner_id: req.user.id } })
         .then(
-            function findSuccess(game) {
+            (game) => {
                 if (!game) {
-                    return res.status(StatusCodes.NOT_FOUND).json({
+                     res.status(StatusCodes.NOT_FOUND).json({
                         message: "Game Not Found"
                     })
+                } else {
+                    res.status(StatusCodes.OK).json({
+                        game
+                    })
                 }
-                res.status(StatusCodes.OK).json({
-                    game: game
-                })
+
             },
 
-            function findFail(err) {
+            () => {
                 res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
                     message: "Internal Server Error"
                 })
@@ -54,14 +56,14 @@ router.post('/create', (req, res) => {
         have_played: req.body.game.have_played
     })
         .then(
-            function createSuccess(game) {
+            (game) => {
                 res.status(StatusCodes.OK).json({
-                    game: game,
+                    game,
                     message: "Game created."
                 })
             },
 
-            function createFail(err) {
+            (err) => {
                 res.status(StatusCodes.INTERNAL_SERVER_ERROR).send(err.message)
             }
         )
@@ -82,14 +84,14 @@ router.put('/update/:id', (req, res) => {
             }
         })
         .then(
-            function updateSuccess(game) {
+            (game) => {
                 res.status(StatusCodes.OK).json({
-                    game: game,
+                    game,
                     message: "Successfully updated."
                 })
             },
 
-            function updateFail(err) {
+            (err) => {
                 res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
                     message: err.message
                 })
@@ -106,14 +108,14 @@ router.delete('/remove/:id', (req, res) => {
         }
     })
     .then(
-        function deleteSuccess(game) {
+        (game) => {
             res.status(StatusCodes.OK).json({
-                game: game,
+                game,
                 message: "Successfully deleted"
             })
         },
 
-        function deleteFail(err) {
+        (err) => {
             res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
                 error: err.message
             })
