+++ src/controllers/usercontroller.js
@@ -1,7 +1,8 @@
-var router = require('express').Router();
-var bcrypt = require('bcryptjs');
-var jwt = require('jsonwebtoken');
-var StatusCodes = require('http-status-codes').StatusCodes;
+const router = require('express').Router();
+const bcrypt = require('bcryptjs');
+const jwt = require('jsonwebtoken');
+const {StatusCodes} = require('http-status-codes');
+const User = require('../models/user');
 
 router.post('/signup', (req, res) => {
     User.create({
@@ -11,15 +12,15 @@ router.post('/signup', (req, res) => {
         email: req.body.user.email,
     })
         .then(
-            function signupSuccess(user) {
-                let token = jwt.sign({ id: user.id }, 'lets_play_sum_games_man', { expiresIn: 60 * 60 * 24 });
+            (user) => {
+                const token = jwt.sign({ id: user.id }, 'lets_play_sum_games_man', { expiresIn: 60 * 60 * 24 });
                 res.status(StatusCodes.OK).json({
-                    user: user,
-                    token: token
+                    user,
+                    token
                 })
             },
 
-            function signupFail(err) {
+            (err) => {
                 res.status(StatusCodes.INTERNAL_SERVER_ERROR).send(err.message)
             }
         )
@@ -28,11 +29,11 @@ router.post('/signup', (req, res) => {
 router.post('/signin', (req, res) => {
     User.findOne({ where: { username: req.body.user.username } }).then(user => {
         if (user) {
-            bcrypt.compare(req.body.user.password, user.passwordHash, function (err, matches) {
+            bcrypt.compare(req.body.user.password, user.passwordHash, (err, matches) => {
                 if (matches) {
-                    var token = jwt.sign({ id: user.id }, 'lets_play_sum_games_man', { expiresIn: 60 * 60 * 24 });
+                    const token = jwt.sign({ id: user.id }, 'lets_play_sum_games_man', { expiresIn: 60 * 60 * 24 });
                     res.json({
-                        user: user,
+                        user,
                         message: "Successfully authenticated.",
                         sessionToken: token
                     });
