+++ src/db.js
@@ -1,6 +1,6 @@
 const Sequelize = require('sequelize');
 require('dotenv').config();
-                                //database username   password
+                                // database username   password
 const sequelize = new Sequelize(process.env.DB, process.env.DB_USER, process.env.DB_PASSWORD, {
     port: process.env.DB_PORT,
     host: process.env.DB_HOST,
@@ -8,11 +8,11 @@ const sequelize = new Sequelize(process.env.DB, process.env.DB_USER, process.env
 })
 
 sequelize.authenticate().then(
-    function success() {
+    () => {
         console.log("Connected to DB");
     },
 
-    function fail(err) {
+    (err) => {
         console.log(`Error: ${err}`);
     }
 )
